1. Fixed multiple bugs in ReccurentUploader 2. Fixed multiple bugs which caused notifications not to appear 3. When fetching parcels from server in ParcelsTableViewController, they are also being persisted in CoreData

+++ PharmaSupplyChain/CDCreatedParcel+CoreDataProperties.swift
@@ -19,7 +19,7 @@ extension CDCreatedParcel {
     
     @NSManaged public var tntNumber: String
     @NSManaged public var sensorMAC: String
-    @NSManaged public var maxFailsTemp: Int
+    @NSManaged public var maxFailsTemp: Int16
     
     /* relationships */
     @NSManaged public var tempCategory: CDTempCategory
